Use variable name instead of id to identify simulation variables. Move brainwaves gripper simulation into its own folder.
