+++ docs/variables.md
@@ -34,7 +34,7 @@ The second use case involves premade simulations, which users access through Sim
 #### Simulations
 Simulations are at the heart of this use case. They are defined in the `src/plugins/simulator/simulation/` directory, and generally describe how a simulation should work. Each simulation contains several properties:
 - delay: The number of milliseconds between steps of the simulation.
-- variables: A list of `Variable` snapshots that will be defined when the simulation is initialized. Each should have an `id`, a `displayName`, and a `value`. A `unit` is optional. These will often also include relevant `labels` (see _Dataflow Tile_ below), as well as an `icon`.
+- variables: A list of `Variable` snapshots that will be defined when the simulation is initialized. Each should have an `name` (which should be unique so the variable can be identified), a `displayName`, and a `value`. A `unit` is optional. These will often also include relevant `labels` (see _Dataflow Tile_ below), as well as an `icon`.
 - values: A dictionary containing values to assign to variables. Each key should be a variable's name, and each value is a list containing values that will be assigned to the variable in sequence as the simulation progresses. Each step, the variable is assigned the next value in its list, and the values are repeated indefinitely while the simulation runs. For example, `values: { "input_test": [0, 1] }` will make the variable with name "input_test" have value 0 on the first step, 1 on the second step, 0 on the third, etc.
 - step: An optional custom function run on each step. It receives a dictionay containing `frame` and `variables` as a parameter. This function can be used to update variables in terms of each other, for example. It is called after values are updated using _values_ as described above.
 - component: A component used to display the state of the simulation in the Simulator Tile. It receives `frame` and `variables` as props. This will ususally involve animations and visualizations that are controlled by the variables.
