+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arduino/arduinoKit_00000.png
