+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arm/armAnimation_00000.png
