+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arm/armAnimation_00001.png
