+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arm/armAnimation_00004.png
