+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arm/armAnimation_00006.png
