+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/arm/armAnimation_00009.png
