+++ src/plugins/simulator/simulations/brainwaves-gripper/assets/gripper/grabberAnimation_withCup_00001.png
