+++ src/plugins/simulator/simulations/brainwaves-gripper/brainwaves-gripper-assets.tsx
