+++ src/plugins/simulator/simulations/simulation-utilities.ts
@@ -1,7 +1,7 @@
 import { VariableType } from "@concord-consortium/diagram-view";
 
-export function findVariable(variableId: string, variables: VariableType[]) {
-  return variables.find(variable => variable.id === variableId);
+export function findVariable(variableName: string, variables: VariableType[]) {
+  return variables.find(variable => variable.name === variableName);
 }
 
 export function getFrame(percent: number, numFrames: number) {
