+++ src/plugins/simulator/simulations/simulations.ts
@@ -1,4 +1,4 @@
-import { brainwavesGripperSimulation, kBrainwavesKey } from "./brainwaves-gripper";
+import { brainwavesGripperSimulation, kBrainwavesKey } from "./brainwaves-gripper/brainwaves-gripper";
 import { ISimulation } from "./simulation-types";
 
 export const simulations: Record<string, ISimulation> = {
